+++ nilearn/datasets/__init__.py
@@ -19,7 +19,7 @@ from .atlas import (fetch_atlas_craddock_2012, fetch_atlas_destrieux_2009,
                     fetch_atlas_yeo_2011, fetch_atlas_aal,
                     fetch_atlas_basc_multiscale_2015,
                     fetch_coords_dosenbach_2010)
-
+from .utils import get_data_dirs
 
 __all__ = ['MNI152_FILE_PATH', 'fetch_icbm152_2009', 'load_mni152_template',
            'fetch_oasis_vbm',
@@ -34,4 +34,5 @@ __all__ = ['MNI152_FILE_PATH', 'fetch_icbm152_2009', 'load_mni152_template',
            'fetch_atlas_yeo_2011', 'fetch_mixed_gambles', 'fetch_atlas_aal',
            'fetch_megatrawls_netmats', 'fetch_cobre',
            'fetch_atlas_basc_multiscale_2015', 'fetch_coords_dosenbach_2010',
-           'load_mni152_brain_mask', 'fetch_icbm152_brain_gm_mask']
+           'load_mni152_brain_mask', 'fetch_icbm152_brain_gm_mask',
+           'get_data_dirs']